refactor: reorganize settings into a 7-tab structure and align i18n keys
- Consolidated the previous 10 tabs into 7 logical tabs (General, Bookshelf, File Navigator, Reader, Rendering, Developer, About) to improve UX.
- Grouped related items within each tab using `ListSubheader` components (e.g., separating "Display & Language" from "Startup Behavior").
- Moved `StartupSettings`, `PageSettings`, `NovelReaderSettings`, and `HistorySettings` items into their newly integrated parent folders (`GeneralSettings`, `ReaderSettings`, `RenderingSettings`).
- Deleted obsolete wrapper components.
- Restructured `en-US.json` and `ja-JP.json` to match the new component hierarchy and updated the i18n keys across the components.
- Updated `SettingsView` and related unit tests to reflect the new structure.
