+++ src/features/Settings/components/GeneralSettings/Items/CheckUpdateOnStartupSetting.test.tsx
