+++ src/features/Settings/components/GeneralSettings/Items/InitialViewSetting.test.tsx
