+++ src/features/Settings/components/ReaderSettings/Items/ShowCoverAsSinglePageSetting.test.tsx
