+++ src/features/Settings/components/RenderingSettings/Items/ImageCacheSizeSetting.test.tsx
