+++ src/features/Settings/components/RenderingSettings/RenderingSettings.tsx
@@ -1,13 +1,15 @@
-import { Divider, List } from "@mui/material";
+import { Divider, List, ListSubheader } from "@mui/material";
 import { useTranslation } from "react-i18next";
 import SettingsPanel from "../SettingsPanel";
+import ImageCacheSizeSetting from "./Items/ImageCacheSizeSetting";
 import ImageResamplingMethodSetting from "./Items/ImageResamplingMethodSetting";
 import MaxImageHeightSetting from "./Items/MaxImageHeightSetting";
 import PdfRenderResolutionHeightSetting from "./Items/PdfRenderResolutionHeightSetting";
+import PreloadPageCountSetting from "./Items/PreloadPageCountSetting";
 import ThumbnailPreviewSetting from "./Items/ThumbnailPreviewSetting";
 
 /**
- * Rendering settings component.
+ * Rendering and performance settings component.
  */
 export default function RenderingSettings() {
   const { t } = useTranslation();
@@ -15,12 +17,23 @@ export default function RenderingSettings() {
   return (
     <SettingsPanel title={t("settings.rendering.title")}>
       <List>
+        <ListSubheader disableSticky color="primary">
+          {t("settings.rendering.headers.rendering")}
+        </ListSubheader>
         <ThumbnailPreviewSetting />
         <Divider />
         <MaxImageHeightSetting />
         <ImageResamplingMethodSetting />
         <Divider />
         <PdfRenderResolutionHeightSetting />
+      </List>
+      <List>
+        <ListSubheader disableSticky color="primary">
+          {t("settings.rendering.headers.cache-and-preload")}
+        </ListSubheader>
+        <PreloadPageCountSetting />
+        <Divider />
+        <ImageCacheSizeSetting />
       </List>
     </SettingsPanel>
   );
