+++ src/i18n/locales/en-US.json
@@ -128,6 +128,10 @@
     "general": {
       "tab-name": "General",
       "title": "General Settings",
+      "headers": {
+        "display-and-language": "Display & Language",
+        "startup-behavior": "Startup Behavior"
+      },
       "language": {
         "title": "Language",
         "english": "English",
@@ -142,11 +146,7 @@
         "system": "System",
         "light": "Light",
         "dark": "Dark"
-      }
-    },
-    "startup": {
-      "tab-name": "Startup",
-      "title": "Startup Settings",
+      },
       "initial-view": {
         "title": "Initial View on Startup",
         "reader": "Reader",
@@ -169,9 +169,23 @@
         "description": "Automatically scrolls long titles in the bookshelf view. Turning this off can improve performance if the bookshelf feels slow."
       }
     },
-    "page": {
-      "tab-name": "Page",
-      "title": "Page Settings",
+    "file-navigator": {
+      "tab-name": "File Navigator",
+      "title": "File Navigator Settings",
+      "home-directory-title": "Home Directory",
+      "dir-watch": {
+        "title": "Watch a directory or file for changes",
+        "warn-message": "Enabling this feature may cause high system load when accessing directories with a large volume of files."
+      }
+    },
+    "reader": {
+      "tab-name": "Reader Settings",
+      "title": "Reader Settings",
+      "headers": {
+        "comic": "Comic",
+        "novel": "Novel",
+        "history": "History"
+      },
       "first-page": {
         "title": "Show the initial page in single-page view"
       },
@@ -188,40 +202,25 @@
         "mouse-back": "Mouse Back Button",
         "mouse-forward": "Mouse Forward Button"
       },
-      "cache": {
-        "title": "Cache Settings",
-        "preload-page-count": {
-          "title": "Preload Page Count",
-          "description": "The number of pages to preload in each direction (forward and backward).",
-          "error-message": "Failed to set preload page count. Please check your input."
-        },
-        "image-cache-size": {
-          "title": "Image Cache Size (MiB)",
-          "description": "The maximum size of the image memory cache in MiB.\nChanging this will clear the current cache.",
-          "error-message": "Failed to set image cache size. Please check your input."
-        }
-      }
-    },
-    "file-navigator": {
-      "tab-name": "File Navigator",
-      "title": "File Navigator Settings",
-      "home-directory-title": "Home Directory",
-      "dir-watch": {
-        "title": "Watch a directory or file for changes",
-        "warn-message": "Enabling this feature may cause high system load when accessing directories with a large volume of files."
-      }
-    },
-    "history": {
-      "tab-name": "History",
-      "title": "History Settings",
-      "feature-toggle": {
+      "font": {
+        "title": "Font Family",
+        "default-font-name": "Default"
+      },
+      "font-size": {
+        "title": "Font Size"
+      },
+      "history-feature": {
         "title": "Reading History",
         "description": "When enabled, reading history will be saved for each book you read."
       }
     },
     "rendering": {
-      "tab-name": "Rendering",
-      "title": "Rendering Settings",
+      "tab-name": "Rendering & Performance",
+      "title": "Rendering & Performance",
+      "headers": {
+        "rendering": "Rendering",
+        "cache-and-preload": "Cache & Preload"
+      },
       "preview": {
         "title": "Show preview while loading",
         "description": "Displays a low-resolution thumbnail until the full image is loaded."
@@ -252,17 +251,19 @@
         "description": "Height of the PDF rendering area in pixels.",
         "range-error-message": "Please enter a positive integer.",
         "error-message": "Failed to set PDF rendering height. Please check your input."
-      }
-    },
-    "novel-reader": {
-      "tab-name": "Novel Reader",
-      "title": "Novel Reader Settings",
-      "font": {
-        "title": "Font Family",
-        "default-font-name": "Default"
       },
-      "font-size": {
-        "title": "Font Size"
+      "cache": {
+        "title": "Cache Settings",
+        "preload-page-count": {
+          "title": "Preload Page Count",
+          "description": "The number of pages to preload in each direction (forward and backward).",
+          "error-message": "Failed to set preload page count. Please check your input."
+        },
+        "image-cache-size": {
+          "title": "Image Cache Size (MiB)",
+          "description": "The maximum size of the image memory cache in MiB.\nChanging this will clear the current cache.",
+          "error-message": "Failed to set image cache size. Please check your input."
+        }
       }
     },
     "developer": {
